Conflicts:
	tools/tolua/cocos2dx.ini

+++ cocos/base/CCNS.cpp
@@ -36,8 +36,8 @@ typedef std::vector<std::string> strArray;
 // string toolkit
 static inline void split(std::string src, const char* token, strArray& vect)
 {
-    int nend=0;
-    int nbegin=0;
+    size_t nend=0;
+    size_t nbegin=0;
     while(nend != -1)
     {
         nend = src.find(token, nbegin);
@@ -65,8 +65,8 @@ static bool splitWithForm(const std::string& str, strArray& strs)
         std::string content = str;
         CC_BREAK_IF(content.length() == 0);
 
-        int nPosLeft  = content.find('{');
-        int nPosRight = content.find('}');
+        size_t nPosLeft  = content.find('{');
+        size_t nPosRight = content.find('}');
 
         // don't have '{' and '}'
         CC_BREAK_IF(nPosLeft == (int)std::string::npos || nPosRight == (int)std::string::npos);
@@ -77,8 +77,8 @@ static bool splitWithForm(const std::string& str, strArray& strs)
         // nothing between '{' and '}'
         CC_BREAK_IF(pointStr.length() == 0);
 
-        int nPos1 = pointStr.find('{');
-        int nPos2 = pointStr.find('}');
+        size_t nPos1 = pointStr.find('{');
+        size_t nPos2 = pointStr.find('}');
         // contain '{' or '}' 
         CC_BREAK_IF(nPos1 != (int)std::string::npos || nPos2 != (int)std::string::npos);
 
@@ -107,8 +107,8 @@ Rect RectFromString(const std::string& str)
         std::string content = str;
 
         // find the first '{' and the third '}'
-        int nPosLeft  = content.find('{');
-        int nPosRight = content.find('}');
+        size_t nPosLeft  = content.find('{');
+        size_t nPosRight = content.find('}');
         for (int i = 1; i < 3; ++i)
         {
             if (nPosRight == (int)std::string::npos)
@@ -120,7 +120,7 @@ Rect RectFromString(const std::string& str)
         CC_BREAK_IF(nPosLeft == (int)std::string::npos || nPosRight == (int)std::string::npos);
 
         content = content.substr(nPosLeft + 1, nPosRight - nPosLeft - 1);
-        int nPointEnd = content.find('}');
+        size_t nPointEnd = content.find('}');
         CC_BREAK_IF(nPointEnd == (int)std::string::npos);
         nPointEnd = content.find(',', nPointEnd);
         CC_BREAK_IF(nPointEnd == (int)std::string::npos);
